openVPN/start_openvpn.bash: removed "sleep 30" and now using ONN719 ping to decide

+++ openVPN/start_openvpn.bash
@@ -7,9 +7,12 @@
 
 script=`realpath $0`
 scriptPath=`dirname $script`
+
+# Variables
 VPNPrefix="moyen"
 primaryKeyDir=${scriptPath}/current
 backupKeyDir=${primaryKeyDir}/BackupKey
+IPTest="10.7.19.254"
 
 primaryKey=`cd $primaryKeyDir;ls *.key`
 primaryKey=`echo $primaryKey | sed "s/.key//g"`
@@ -43,35 +46,56 @@ then
 	echo "  First deleting old $primaryKeyDir/nohup.out"
 	sudo rm -f $primaryKeyDir/nohup.out
 	cd $primaryKeyDir; sudo nohup ./start_openvpn.sh & 2>/dev/null
+
 	echo "------------------------------------------------"
-	echo "... Waiting 30s so we have time to connect ..."
+	echo "... Giving a max of 30s to connect ..."
 	echo "... and see if we need to start $backupKey ..."
-	sleep 30 # waiting for the first key to try and connect
-
+	Delay=0
+	while [ $Delay -ne 30 ]
+	do
+		if sudo grep "Initialization Sequence Completed" $primaryKeyDir/nohup.out >/dev/null 2>&1
+		then
+			Delay=30
+		else
+			((Delay++))
+		fi
+		sleep 1
+	done
+			
 	# Checking if first key is connected and if no trying the second one
-	ps -ef | grep -v grep | grep $primaryKey > /dev/null
-        running=`echo $?`
-
-        if [ $running -ne 0 ]
+	if ! ping -c1 $IPTest >/dev/null
 	then
 		echo
 		echo "------------------------------------------------"
-		echo "$primaryKey failed. Trying to start openVPN $backupKey"
+		echo "Cannot access to $IPTest..."
+		echo "  --> $primaryKey failed. Trying to start openVPN $backupKey"
 		echo "  First deleting old $backupKeyDir/nohup.out"
 		sudo rm -f $backupKeyDir/nohup.out
 		cd $backupKeyDir; sudo nohup ./start_openvpn.sh & 2>/dev/null
-		sleep 30 # waiting for the backup key to try and connect
-		ps -ef | grep -v grep | grep $backupKey > /dev/null
-        	backupActive=`echo $?`
-		if [ $backupActive -eq 0 ]
+
+		Delay=0
+		while [ $Delay -ne 30 ]
+		do
+			if sudo grep "Initialization Sequence Completed" $backupKeyDir/nohup.out >/dev/null 2>&1
+			then
+				Delay=30
+			else
+				((Delay++))
+			fi
+			sleep 1
+		done
+
+		if ping -c1 $IPTest >/dev/null
 		then
 			echo
 			echo "------------------------------------------------"
-			echo "$backupKey running. Exiting..."
+			echo "$IPTest is responding"
+			echo "  --> $backupKey running. Exiting..."
 		else
 			echo
 			echo "------------------------------------------------"
-			echo "$backupKey FAILED !!! Needs repair..."
+			echo "Cannot access to $IPTest..."
+			echo "  --> $backupKey FAILED !!! Needs repair..."
 			echo "Exiting..."
 			exit 1
 		fi
@@ -79,7 +103,8 @@ then
 	else
 		echo
 		echo "------------------------------------------------"
-		echo "$primaryKey running. Exiting..."
+		echo "$IPTest is responding"
+		echo "  --> $primaryKey running. Exiting..."
 	fi
 else
 	echo
@@ -88,7 +113,7 @@ else
 fi
 
 echo
-pingTest=`ping -c 1 10.7.19.254 | grep "packets transmitted"`
-echo "ping test ONN719 -> $pingTest"
+pingTest=`ping -c 1 $IPTest | grep "packets transmitted"`
+echo "ping test $IPTest -> $pingTest"
 
 echo
